pyproject.toml
[project]
name = "prueba_abpo"
version = "0.1.0"
description = ""
authors = [
    {name = "pablito Cortes",email = "cjuan1597@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.121.1,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.0,<10.0.0)"
]
